+++ src/api/projects/repository/files/delete.rs
@@ -33,6 +33,11 @@ pub struct DeleteFile<'a> {
     /// The name of the author for the new commit.
     #[builder(setter(into), default)]
     author_name: Option<Cow<'a, str>>,
+    /// The last commit ID the path was modified in.
+    ///
+    /// Used to detect conflicts.
+    #[builder(setter(into), default)]
+    last_commit_id: Option<Cow<'a, str>>,
 }
 
 impl<'a> DeleteFile<'a> {
@@ -64,7 +69,8 @@ impl<'a> Endpoint for DeleteFile<'a> {
             .push("commit_message", &self.commit_message)
             .push_opt("start_branch", self.start_branch.as_ref())
             .push_opt("author_email", self.author_email.as_ref())
-            .push_opt("author_name", self.author_name.as_ref());
+            .push_opt("author_name", self.author_name.as_ref())
+            .push_opt("last_commit_id", self.last_commit_id.as_ref());
 
         params.into_body()
     }
@@ -233,6 +239,32 @@ mod tests {
             .branch("branch")
             .commit_message("commit message")
             .author_name("Arthur Developer")
+            .build()
+            .unwrap();
+        api::ignore(endpoint).query(&client).unwrap();
+    }
+
+    #[test]
+    fn endpoint_last_commit_id() {
+        let endpoint = ExpectedUrl::builder()
+            .method(Method::DELETE)
+            .endpoint("projects/simple%2Fproject/repository/files/path%2Fto%2Ffile")
+            .content_type("application/x-www-form-urlencoded")
+            .body_str(concat!(
+                "branch=branch",
+                "&commit_message=commit+message",
+                "&last_commit_id=deadbeefdeadbeefdeadbeefdeadbeefdeadbeef",
+            ))
+            .build()
+            .unwrap();
+        let client = SingleTestClient::new_raw(endpoint, "");
+
+        let endpoint = DeleteFile::builder()
+            .project("simple/project")
+            .file_path("path/to/file")
+            .branch("branch")
+            .commit_message("commit message")
+            .last_commit_id("deadbeefdeadbeefdeadbeefdeadbeefdeadbeef")
             .build()
             .unwrap();
         api::ignore(endpoint).query(&client).unwrap();