+++ CHANGELOG.md
@@ -30,6 +30,8 @@
   * Support including HTML descriptions in project releases.
   * Support sorting parameters for project releases.
   * Support including trailers in commit queries.
+  * Support `execute_filemode` and `last_commit_id` when editing files through
+    the API.
 
 ## Deprecations